Collision sound bug fix + general organization
Fixed mass collision sound issue by creating TemporarySfxMute.cs (added to each playable scene via prefab called Mute Script), which mutes all SFX except music and ambience for one second on scene load.

Also organized audio assets (scripts/prefabs) into folders.
