+++ Cooking VR/Assets/Prefabs/Audio/Audio -Environ..prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 07bf04a33a6d0a94eb918658506160d8
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
