+++ Cooking VR/Assets/Scripts/Audio/VoTriggers.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 894329f8ccce00e419290ade6a86cfc8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
