--- Cooking VR/Assets/Scripts/TutorialVO.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c3250238b5bf87244afba4dcffc6f30e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
