--- Cooking VR/Assets/Scripts/VoTriggers.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 894329f8ccce00e419290ade6a86cfc8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
