+++ Cooking VR/Assets/StreamingAssets/Audio/GeneratedSoundBanks/Wwise_IDs.h
@@ -109,6 +109,7 @@ namespace AK
     namespace GAME_PARAMETERS
     {
         static const AkUniqueID GAMEPARAMETERINTENSITY = 964045659U;
+        static const AkUniqueID SFX_VOLUME = 1564184899U;
     } // namespace GAME_PARAMETERS
 
     namespace BANKS
@@ -120,6 +121,7 @@ namespace AK
     namespace BUSSES
     {
         static const AkUniqueID MASTER_AUDIO_BUS = 3803692087U;
+        static const AkUniqueID SFX = 393239870U;
     } // namespace BUSSES
 
     namespace AUDIO_DEVICES
