+++ Cooking VR/Cooking VR_WwiseProject/Cooking VR_WwiseProject.Blue.validationcache
@@ -11,20 +11,20 @@
 	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Presets\Factory Spatial Audio.wwu" DateLow="1590038476" DateHigh="0"/>
 	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Presets\Default Work Unit.wwu" DateLow="1590038476" DateHigh="0"/>
 	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Modulators\Default Work Unit.wwu" DateLow="1590038474" DateHigh="0"/>
-	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Mixing Sessions\Default Work Unit.wwu" DateLow="1590038474" DateHigh="0"/>
-	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Master-Mixer Hierarchy\Default Work Unit.wwu" DateLow="1590038474" DateHigh="0"/>
+	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Mixing Sessions\Default Work Unit.wwu" DateLow="1593286005" DateHigh="0"/>
+	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Master-Mixer Hierarchy\Default Work Unit.wwu" DateLow="1593300828" DateHigh="0"/>
 	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Interactive Music Hierarchy\Default Work Unit.wwu" DateLow="1592075715" DateHigh="0"/>
-	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Game Parameters\Default Work Unit.wwu" DateLow="1590038473" DateHigh="0"/>
-	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Events\Default Work Unit.wwu" DateLow="1592683424" DateHigh="0"/>
+	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Game Parameters\Default Work Unit.wwu" DateLow="1593287310" DateHigh="0"/>
+	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Events\Default Work Unit.wwu" DateLow="1593286005" DateHigh="0"/>
 	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Effects\Factory Effects.wwu" DateLow="1590038473" DateHigh="0"/>
 	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Effects\Default Work Unit.wwu" DateLow="1590038473" DateHigh="0"/>
 	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Dynamic Dialogue\Default Work Unit.wwu" DateLow="1590038473" DateHigh="0"/>
 	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Cooking VR_WwiseProject.wproj" DateLow="1590038473" DateHigh="0"/>
-	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Cooking VR_WwiseProject.Blue.wsettings" DateLow="1592680542" DateHigh="0"/>
+	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Cooking VR_WwiseProject.Blue.wsettings" DateLow="1593281862" DateHigh="0"/>
 	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Conversion Settings\Factory Conversion Settings.wwu" DateLow="1590038473" DateHigh="0"/>
 	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Conversion Settings\Default Work Unit.wwu" DateLow="1590038473" DateHigh="0"/>
 	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Control Surface Sessions\Default Work Unit.wwu" DateLow="1590038473" DateHigh="0"/>
 	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Audio Devices\Default Work Unit.wwu" DateLow="1590038473" DateHigh="0"/>
 	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Attenuations\Default Work Unit.wwu" DateLow="1591475069" DateHigh="0"/>
-	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Actor-Mixer Hierarchy\Default Work Unit.wwu" DateLow="1592683757" DateHigh="0"/>
+	<ValidationCacheItem File="D:\Documents\GitHub\Cooking_VR\Cooking VR\Cooking VR_WwiseProject\Actor-Mixer Hierarchy\Default Work Unit.wwu" DateLow="1593301022" DateHigh="0"/>
 </ValidationCache>
