+++ Cooking VR/Cooking VR_WwiseProject/Cooking VR_WwiseProject.Blue.wsettings
@@ -55,12 +55,13 @@
 				<Property Name="MeterView\MeterMode1Viewpoint1" Type="int32" Value="2"/>
 				<Property Name="MeterView\MeterMode1Viewpoint2" Type="int32" Value="3"/>
 				<Property Name="MeterView\MeterMode1Viewpoint3" Type="int32" Value="0"/>
+				<Property Name="Mixing Desk\GroupCollapse\x64_{89394D22-7EF0-4ECD-B1FD-26D6CF633DF6}" Type="string" Value=""/>
 				<Property Name="Mixing Desk\LastSession" Type="string" Value="{00000000-0000-0000-0000-000000000000}"/>
 				<Property Name="MonitoringManager\GameObjectFilter\GameObjectFilterVersion" Type="int32" Value="1"/>
 				<Property Name="MonitoringManager\GameObjectFilter\x64_GameObjectFilters" Type="string" Value=""/>
 				<Property Name="ProjectExplorer\FirstVisible_10_1" Type="string" Value="{4C96B38B-3DB4-40C7-91C9-665510E38993}"/>
 				<Property Name="ProjectExplorer\FirstVisible_11_1" Type="string" Value="{23EC2DE3-3ED8-4B82-AC99-92E25F2E4C6B}"/>
-				<Property Name="ProjectExplorer\FirstVisible_1_1" Type="string" Value="{C3A2E2F1-E655-4574-A195-57E502BE748B}"/>
+				<Property Name="ProjectExplorer\FirstVisible_1_1" Type="string" Value="{F885AA4B-4AC3-4CD1-A575-1F7C7EA1AAB8}"/>
 				<Property Name="ProjectExplorer\FirstVisible_3_1" Type="string" Value="{25BF5F49-CBE0-49DF-B4AF-F1C6A326EE7C}"/>
 				<Property Name="ProjectExplorer\FirstVisible_4_1" Type="string" Value="{CB6175F1-6080-4A23-92EF-0807E304D0C1}"/>
 				<Property Name="ProjectExplorer\FirstVisible_6_1" Type="string" Value="{C98E673F-A24C-48A8-960F-10565FFE70A0}"/>
@@ -71,21 +72,22 @@
 				<Property Name="ProjectExplorer\x64_Expanded_13_1" Type="string" Value=""/>
 				<Property Name="ProjectExplorer\x64_Expanded_14_1" Type="string" Value=""/>
 				<Property Name="ProjectExplorer\x64_Expanded_15_1" Type="string" Value=""/>
-				<Property Name="ProjectExplorer\x64_Expanded_16_1" Type="string" Value=""/>
+				<Property Name="ProjectExplorer\x64_Expanded_16_1" Type="string" Value="k8xWogKdsUCffloPd/naiw=="/>
 				<Property Name="ProjectExplorer\x64_Expanded_17_1" Type="string" Value=""/>
 				<Property Name="ProjectExplorer\x64_Expanded_18_1" Type="string" Value=""/>
 				<Property Name="ProjectExplorer\x64_Expanded_19_1" Type="string" Value=""/>
-				<Property Name="ProjectExplorer\x64_Expanded_1_1" Type="string" Value=""/>
+				<Property Name="ProjectExplorer\x64_Expanded_1_1" Type="string" Value="2KQUFaYdKkGhfnXKDCFJ86F9qzzSqolLqqzJ4flKwkw="/>
 				<Property Name="ProjectExplorer\x64_Expanded_20_1" Type="string" Value=""/>
 				<Property Name="ProjectExplorer\x64_Expanded_21_1" Type="string" Value=""/>
-				<Property Name="ProjectExplorer\x64_Expanded_2_1" Type="string" Value="TBetEVGsM0qZBXh6G8/1YD9vGHdhX5dPlxevtUAqMU8="/>
+				<Property Name="ProjectExplorer\x64_Expanded_2_1" Type="string" Value="P28Yd2Ffl0+XF6+1QCoxTw=="/>
 				<Property Name="ProjectExplorer\x64_Expanded_3_1" Type="string" Value="BJpAlfEDxEO/QYmNmXvdVw=="/>
-				<Property Name="ProjectExplorer\x64_Expanded_4_1" Type="string" Value=""/>
+				<Property Name="ProjectExplorer\x64_Expanded_4_1" Type="string" Value="DRxTr5Ec80SvLVzd0gYDF/jpZh7CF51IqylDDsUpFgM="/>
 				<Property Name="ProjectExplorer\x64_Expanded_5_1" Type="string" Value=""/>
 				<Property Name="ProjectExplorer\x64_Expanded_6_1" Type="string" Value="lAb1JI0l3UqzMWadlR8NrA=="/>
-				<Property Name="ProjectExplorer\x64_Expanded_7_1" Type="string" Value=""/>
+				<Property Name="ProjectExplorer\x64_Expanded_7_1" Type="string" Value="PWNj1c5goEGt4rbgnS3/Ng=="/>
 				<Property Name="ProjectExplorer\x64_Expanded_8_1" Type="string" Value=""/>
 				<Property Name="RTPC\GameParameterSimValue\{10CAFFE5-0DBC-41A8-B9F3-FC1DF08F2E04}" Type="Real64" Value="1"/>
+				<Property Name="RTPC\GameParameterSimValue\{E1C39B7E-E48D-4E85-88B3-A0CD43584F73}" Type="Real64" Value="8.86075949367088"/>
 				<Property Name="SoundBank Manager\x64_Expand States" Type="string" Value="lAb1JI0l3UqzMWadlR8NrA=="/>
 				<Property Name="SoundBankHeaderFilePath" Type="string" Value="GeneratedSoundBanks\"/>
 				<Property Name="SoundBankPaths" Type="string">
