+++ Cooking VR/Cooking VR_WwiseProject/Game Parameters/Default Work Unit.wwu
@@ -10,6 +10,7 @@
 						<Property Name="Min" Type="Real64" Value="1"/>
 					</PropertyList>
 				</GameParameter>
+				<GameParameter Name="SFX_Volume" ID="{E1C39B7E-E48D-4E85-88B3-A0CD43584F73}"/>
 			</ChildrenList>
 		</WorkUnit>
 	</GameParameters>
