+++ Cooking VR/Cooking VR_WwiseProject/GeneratedSoundBanks/Windows/SoundbanksInfo.xml
@@ -15,6 +15,9 @@
 			<ObjectPath>Init</ObjectPath>
 			<ShortName>Init</ShortName>
 			<Path>Init.bnk</Path>
+			<GameParameters>
+				<GameParameter Id="1564184899" Name="SFX_Volume" ObjectPath="\Game Parameters\Default Work Unit\SFX_Volume"/>
+			</GameParameters>
 		</SoundBank>
 		<SoundBank Id="3989171651" Language="SFX">
 			<ObjectPath>\SoundBanks\Default Work Unit\Main_Bank</ObjectPath>
